refactor testing; add README
